LayoutTests/fast/forms/ios/click-should-not-suppress-misspelling-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS input.selectionStart is 0
PASS input.selectionEnd is 7
